+++ MCP3424.cpp
@@ -24,8 +24,9 @@ MCP3424::MCP3424(uint8_t address, TwoWire *wire)
   _gain = 1;
   _bits = 12;
   _config = 0x10;  //  default
+  _lastRead = 0;
+  _raw = 0;
 }
-
 
 bool MCP3424::begin()
 {
@@ -41,29 +42,41 @@ bool MCP3424::begin()
   return false;
 }
 
-
 bool MCP3424::isConnected()
 {
   _wire->beginTransmission(_address);
   return ( _wire->endTransmission() == 0);
 }
 
-
 uint8_t MCP3424::getAddress()
 {
   return _address;
 }
-
 
 uint8_t MCP3424::getMaxChannels()
 {
   return _maxChannels;
 }
 
-
 int32_t MCP3424::read()
 {
   return readRaw();
+}
+
+void MCP3424::requestSingleShot()
+{
+  setSingleShotMode();
+}
+
+bool MCP3424::isReady()
+{
+  readRaw();
+  return (_config & 0x80) == 0x00;
+}
+
+uint32_t MCP3424::lastRead()
+{
+  return _lastRead;
 }
 
 
@@ -79,12 +92,10 @@ float MCP3424::readVolts()
   return read() * (15.625e-6 * (1L << (18 - _bits))) / _gain;
 }
 
-
 float MCP3424::readMilliVolts()
 {
   return read() * (15.625e-3 * (1L << (18 - _bits))) / _gain;
 }
-
 
 float MCP3424::readMicroVolts()
 {
@@ -111,7 +122,6 @@ bool MCP3424::setChannel(uint8_t channel)
   writeConfig();
   return true;
 }
-
 
 uint8_t MCP3424::getChannel()
 {
@@ -140,7 +150,6 @@ bool MCP3424::setGain(uint8_t gain)
   return true;
 }
 
-
 uint8_t MCP3424::getGain()
 {
   return _gain;
@@ -166,7 +175,6 @@ bool MCP3424::setResolution(uint8_t bits)
   return true;
 }
 
-
 uint8_t MCP3424::getResolution()
 {
   //  return 12 + 2 * ((_config >> 2) & 0x03);
@@ -181,7 +189,6 @@ uint16_t MCP3424::getConversionDelay()
   return _interval[idx];
 }
 
-
 void MCP3424::setContinuousMode()
 {
   if (getMode() != 1)
@@ -194,11 +201,9 @@ void MCP3424::setContinuousMode()
 
 void MCP3424::setSingleShotMode()
 {
-  if (getMode() != 0)
-  {
-    _config &= ~0x10;
-    writeConfig();
-  }
+  _config &= ~0x10;  //  set single shot mode
+  _config |= 0x80;   //  and trigger single shot
+  writeConfig();
 }
 
 
@@ -228,8 +233,8 @@ int32_t MCP3424::readRaw()
 {
   int32_t rv = 0;
 
-  uint8_t bitsToRead = 2;
-  if (_bits == 18) bitsToRead = 3;
+  uint8_t bitsToRead = 3;
+  if (_bits == 18) bitsToRead = 4;
   uint32_t signBit = (1UL << (_bits -1));
   uint32_t extendSign = ~(signBit -1);
 
@@ -238,7 +243,8 @@ int32_t MCP3424::readRaw()
     //  _error = MCP3424_I2C_ERROR;
     return 0;
   }
-  if (bitsToRead == 3)
+  //  read data bytes
+  if (bitsToRead == 4)
   {
     rv += _wire->read();
     rv <<= 8;
@@ -247,12 +253,16 @@ int32_t MCP3424::readRaw()
   rv <<= 8;
   rv += _wire->read();
 
+  //  read configuration byte
+  _config = _wire->read();
   //  handle sign bit.
   if (rv & signBit)
   {
     //  sign extend 14 bits
     rv |= extendSign;
   }
+  _lastRead = millis();
+  _raw = rv;
   return rv;
 }
 */
@@ -266,34 +276,42 @@ int32_t MCP3424::readRaw()
   //  not optimized compacted code.
   if (_bits == 18)
   {
-    if (_wire->requestFrom(_address, (uint8_t)3) != 3)
+    if (_wire->requestFrom(_address, (uint8_t)4) != 4)
     {
       //  _error = MCP3424_I2C_ERROR;
       return 0;
     }
+    //  read data bytes
     rv += _wire->read();
     rv <<= 8;
     rv += _wire->read();
     rv <<= 8;
     rv += _wire->read();
+    //  read the configuration byte
+    _config = _wire->read();
     //  handle sign bit.
     if (rv & 0x00020000)
     {
       //  sign extend 14 bits
       rv |= 0xFFFC0000;
     }
+    _lastRead = millis();
+    _raw = rv;
     return rv;
   }
 
   //  12, 14, 16 bit
-  if (_wire->requestFrom(_address, (uint8_t)2) != 2)
+  if (_wire->requestFrom(_address, (uint8_t)3) != 3)
   {
     //  _error = MCP3424_I2C_ERROR;
     return 0;
   }
+  //  read data bytes
   rv += _wire->read();
   rv <<= 8;
   rv += _wire->read();
+  //  read the configuration byte
+  _config = _wire->read();
   //  handle sign bit
   if ((_bits == 12) && (rv & 0x0800))
   {
@@ -310,6 +328,8 @@ int32_t MCP3424::readRaw()
     //  sign extend 16 bits
     rv |= 0xFFFF0000;
   }
+  _lastRead = millis();
+  _raw = rv;
   return rv;
 }